Creating CalIPE repository
Toolbox to calibrate the following intensity prediction equation (IPE):

I = C1 + C2.Mag + BETA.log10(Hypo) + GAMMA.Hypo
with C1, C2, BETA and GAMMA the coefficients to calibrate.
I is the predicted intensity, Mag is the magnitude and Hypo is the hypocentral
distance associated to I.
In this initial version, it is possible to calibrate the geometric attenuation
coefficient BETA using Kovesligethy mathematical formulation.